+++ data/duke.txt
@@ -1,3 +1,5 @@
 1.T|Done a 
 2.T|Done a 
-3.E|NotDone a (sk: 10 Oct 2020)
+3.T|Done a 
+4.T|Done a 
+5.E|NotDone a (aa: s)
